requirements.txt
pydantic==2.11.3
python-telegram-bot==22.1
fastapi==0.115.12
uvicorn==0.34.0
openai==0.28.0
python-dotenv==1.1.0
requests==2.32.3